Refactor handling user input into Parser class
